requirements.txt
# local package
-e .

# external requirements
black
ipython
ipykernel
matplotlib
seaborn
pandas
numpy
sklearn
click
Sphinx
coverage
gdown
awscli
python-dotenv>=0.5.1
